+++ src/Console/ConsoleDependencyProvider.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Gacela\Console;
 
 use Gacela\Console\Domain\FilenameSanitizer\FilenameSanitizer;
-use Gacela\Console\Infrastructure\Command\ClearCacheCommand;
 use Gacela\Console\Infrastructure\Command\MakeFileCommand;
 use Gacela\Console\Infrastructure\Command\MakeModuleCommand;
 use Gacela\Framework\AbstractDependencyProvider;
@@ -31,7 +30,6 @@ final class ConsoleDependencyProvider extends AbstractDependencyProvider
         $container->set(self::COMMANDS, static fn () => [
             new MakeFileCommand(),
             new MakeModuleCommand(),
-            new ClearCacheCommand(),
         ]);
     }