+++ src/Console/ConsoleFactory.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace Gacela\Console;
 
-use Gacela\Console\Domain\Cache\CacheClearer;
-use Gacela\Console\Domain\Cache\DirectoryIoInterface;
 use Gacela\Console\Domain\CommandArguments\CommandArgumentsParser;
 use Gacela\Console\Domain\CommandArguments\CommandArgumentsParserInterface;
 use Gacela\Console\Domain\FileContent\FileContentGenerator;
@@ -13,7 +11,6 @@ use Gacela\Console\Domain\FileContent\FileContentGeneratorInterface;
 use Gacela\Console\Domain\FileContent\FileContentIoInterface;
 use Gacela\Console\Domain\FilenameSanitizer\FilenameSanitizer;
 use Gacela\Console\Domain\FilenameSanitizer\FilenameSanitizerInterface;
-use Gacela\Console\Infrastructure\DirectoryIo;
 use Gacela\Console\Infrastructure\FileContentIo;
 use Gacela\Framework\AbstractFactory;
 use Symfony\Component\Console\Command\Command;
@@ -53,14 +50,6 @@ final class ConsoleFactory extends AbstractFactory
         );
     }
 
-    public function createCacheClearer(): CacheClearer
-    {
-        return new CacheClearer(
-            $this->getConfig()->getCacheDir(),
-            $this->createDirectoryIo()
-        );
-    }
-
     private function createFileContentIo(): FileContentIoInterface
     {
         return new FileContentIo();
@@ -74,10 +63,5 @@ final class ConsoleFactory extends AbstractFactory
     private function getTemplateByFilenameMap(): array
     {
         return (array)$this->getProvidedDependency(ConsoleDependencyProvider::TEMPLATE_BY_FILENAME_MAP);
-    }
-
-    private function createDirectoryIo(): DirectoryIoInterface
-    {
-        return new DirectoryIo();
     }
 }