+++ tests/Benchmark/Framework/ClassResolver/FileProfiler/ModuleD/Repository.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Benchmark\Framework\ClassResolver\FileProfiler\ModuleD;
+
+final class Repository
+{
+    public function getAll(): array
+    {
+        return ['d'];
+    }
+}