+++ tests/Feature/Framework/CustomNoCacheDirectory/NoFileCacheFeatureTest.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace GacelaTest\Feature\Framework\CustomNoCacheDirectory;
 
 use Gacela\Framework\Bootstrap\GacelaConfig;
-use Gacela\Framework\ClassResolver\ClassNameCache;
-use Gacela\Framework\ClassResolver\DocBlockService\CustomServicesCache;
+use Gacela\Framework\ClassResolver\ClassNameJsonProfiler;
+use Gacela\Framework\ClassResolver\DocBlockService\CustomServicesJsonProfiler;
 use Gacela\Framework\Gacela;
 use PHPUnit\Framework\TestCase;
 
@@ -15,8 +15,8 @@ final class NoFileCacheFeatureTest extends TestCase
     public function setUp(): void
     {
         Gacela::bootstrap(__DIR__, static function (GacelaConfig $config): void {
-            $config->setCacheEnabled(false);
-            $config->setCacheDirectory('custom/no-caching-dir');
+            $config->setProfilerEnabled(false);
+            $config->setProfilerDirectory('custom/no-profiler-dir');
         });
     }
 
@@ -25,7 +25,7 @@ final class NoFileCacheFeatureTest extends TestCase
         $facade = new Module\Facade();
         self::assertSame('name', $facade->getName());
 
-        self::assertFileDoesNotExist(__DIR__ . '/custom/no-caching-dir/' . ClassNameCache::CACHE_FILENAME);
-        self::assertFileDoesNotExist(__DIR__ . '/custom/no-caching-dir/' . CustomServicesCache::CACHE_FILENAME);
+        self::assertFileDoesNotExist(__DIR__ . '/custom/no-profiler-dir/' . ClassNameJsonProfiler::FILENAME);
+        self::assertFileDoesNotExist(__DIR__ . '/custom/no-profiler-dir/' . CustomServicesJsonProfiler::FILENAME);
     }
 }