+++ tests/Feature/Framework/CustomProfilerDirectory/Module/Persistence/FakeRepository.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Feature\Framework\CustomProfilerDirectory\Module\Persistence;
+
+final class FakeRepository
+{
+    public function findName(): string
+    {
+        return 'name';
+    }
+}