+++ tests/Integration/Framework/DocBlockResolverAware/DocBlockResolverCustomServicesAwareTest.php
@@ -5,7 +5,8 @@ declare(strict_types=1);
 namespace GacelaTest\Integration\Framework\DocBlockResolverAware;
 
 use Gacela\Framework\Bootstrap\GacelaConfig;
-use Gacela\Framework\ClassResolver\DocBlockService\CustomServicesCache;
+use Gacela\Framework\ClassResolver\DocBlockService\CustomServicesJsonProfiler;
+use Gacela\Framework\ClassResolver\InMemoryClassNameCache;
 use Gacela\Framework\Gacela;
 use PHPUnit\Framework\TestCase;
 
@@ -13,7 +14,7 @@ final class DocBlockResolverCustomServicesAwareTest extends TestCase
 {
     public static function setUpBeforeClass(): void
     {
-        CustomServicesCache::resetCache();
+        InMemoryClassNameCache::resetCache();
     }
 
     protected function setUp(): void
@@ -28,7 +29,7 @@ final class DocBlockResolverCustomServicesAwareTest extends TestCase
         $dummy = new DummyDocBlockResolverAware();
         $actual = $dummy->getRepository()->findName();
 
-        self::assertCount(1, CustomServicesCache::getAll());
+        self::assertCount(1, InMemoryClassNameCache::getAllFromKey(CustomServicesJsonProfiler::class));
         self::assertSame('name', $actual);
     }
 
@@ -37,13 +38,13 @@ final class DocBlockResolverCustomServicesAwareTest extends TestCase
      */
     public function test_existing_service_cached(): void
     {
-        self::assertCount(1, CustomServicesCache::getAll());
+        self::assertCount(1, InMemoryClassNameCache::getAllFromKey(CustomServicesJsonProfiler::class));
 
         $dummy = new DummyDocBlockResolverAware();
         $dummy->getRepository()->findName();
         $dummy->getRepository()->findName();
 
-        self::assertCount(1, CustomServicesCache::getAll());
+        self::assertCount(1, InMemoryClassNameCache::getAllFromKey(CustomServicesJsonProfiler::class));
     }
 
     public function test_non_existing_service(): void